Use active pattern matching for event types
Previously, our pattern matching on the incoming data was hacky.

We used an Interface that had a single member to a string enum to do runtime matching. We also cast incoming data to a type before proving it was that type.

Instead use [active pattern matching](https://fsharpforfunandprofit.com/posts/convenience-active-patterns/) to match on the incoming record at runtime.

This allows us to work with an unknown `obj`, and only cast once we determine the type is what we expect.

  - Remove lerna.
  - Update tests.

Signed-off-by: Joe Grund <joe.grund@intel.com>