Update device-scanner events + integrate new test framework
Currently the device-scanner may fail on some uevents.

This is because there are some env variables that are optional
but are not currently marked as such by the type-system.

In particular, `/devices/virtual/block/dm-0` is causing an
exception.

patch up the events so the ones that may not be there are
reflected as `option`.

Longer term, we should be streaming udev db parsing +
piping to `lineDelimitedJSONStream` which should then
parse to a `result` type to ensure we handle faiure.
This will also ensure we don't overflow a local buffer
on larger systems.

In addition, our tests are using mutable bindings for
reusable setup code. This is common in non-functional langs,
but for F# we should be trying to avoid mutation for better
clarity.

Implement the expecto like test-list that landed in fable-jest
recently. This will allow us to remove all mutation in testing.