Update device-scanner rules
We are currently only adding the `IML_SIZE` key
during an add or remove event.

This is fine for reboots, but it doesn't give us
a way to trigger them without retiggering an `add`
event, which we don't want to do.

Refactor the device-scanner.rules file so
it starts with a predicate condition and triggers
`IML_SIZE` on `add|change`.

In addition, IML block_devices.py reads page 80 83 via
`scsi_id`. There is no 100% analouge for that in
any udev rule, so add in the same commands
conditionally on `add|change`.

Signed-off-by: Joe Grund <joe.grund@intel.com>