Remove manual udev parsing
We currently parse the output of `udevadm info -e` into a list of JSON objects.

It turns out we don't need to do this. If we add the change event to what runs the block device listener, we can get the db synced with a `udevadm trigger --action=change --subsystem-match=block`.