Integrate libzfs into device-scanner.
This patch integrates node-libzfs into the device-scanner.
It is a large refactor, as it became apparent that libzfs
will be needed for many different match variants and to
do so cleanly, we should remove the existing active pattern
matching scheme and replace it with a CQRS implementation.

By doing so, we can remove most mutable fields and use something
similar to the elm architecture / redux to emit changes over time
(using scan).

  - Add mock-build.sh for general building with mock.
  - Create spec file for version 2.
  - Refactor Vagrantfile to do a mockbuild and install on provision.
  - Update all copyright years.
  - Refactor listening method to use discrete zedlets and
    udev listeners. By doing this, we can remove the active pattern
    matching.
  - Integrate libzfs during the command phase of CQRS.

Signed-off-by: Joe Grund <joe.grund@intel.com>