Don't assume pool create has occurred on the same node as import
There is an assumption in device-scanner that a pool / dataset
was created on the same node it's imported on.

In an HA scenario, the pool may have been created on an alternate node,
but then failed over to another node.

We should try to update the map in memory, but on a miss
we should pull the data from libzfs.

Signed-off-by: Joe Grund <joe.grund@intel.com>