Enable building in Copr
We are using Copr for production modules.
The approach we are taking is to use rpkg to
build the modules from source, and then build the rpms.

Due to this, we need to make sure we can build the modules
from just the source-tree. This means we need to flesh out
the build step to account for compiling Fable -> JS output.

Signed-off-by: Joe Grund <joe.grund@intel.com>