Alter triggerin scriptlet for ZFS
Alter the scriptlet code that loads the zfs module and starts the zed service.

We will make it a hard requirement that the ZFS kernel
module is loaded and ZED is running prior to installing this
module.

We will do that by installing a service in lustre-ldiskfs-zfs
that does a modprobe.

We will also enforce in monitor mode by enabling ZED in the
install script.

Signed-off-by: Joe Grund <joe.grund@intel.com>