udev serializer
The integration tests will be working with a mixture of data. Most of
the data will be invariant; However, some of the data is variable and
will change on each run. Multipath devices are a perfect example of
this, as well as the root and swap partitions, which are mapped
according to the device mapper. To handle variable data, a serializer
will be used to transform incoming data. Regular expressions will be
used to match the keys on the udev object and the variable portions of
the data will then be transformed such that the data will match the
corresponding snapshot.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>