Cache transmit options in scanner-proxy.
We are currently doing multiple synchronous
disk reads to gather the URL and certs for
communication with the manager.

We are performing these reads every time a request starts,
which will block the main thread.

Instead, we can perform the reads once when the process starts
up, and cache it in memory; as the values are unlikely to change.

Signed-off-by: Joe Grund <joe.grund@intel.com>