Add support for detecting swap devices.
`findmnt --poll` does not track swap mounts in /etc/fstab.
In addition, it's not possible to poll fstab with `findmnt --poll`.
Use `findmnt -P -s -t swap` with a timer unit to detect swap devices.

This is a first pass that only tracks adds and modifications.
We will need to add a more robust implementation in a future pass
that tracks removals as well, possibly by storing the previous tick
and diffing with the current tick.

  - Add swap-emitter.{service,timer}
  - Bump deps to latest
  - Fixup some spec issues.
  - Change preset precedence from 99 to 00 so they all get picked up
early.
  - Fix integration tests
  - Add test for swap.

Signed-off-by: Joe Grund <joe.grund@intel.com>