Use module-tools to build
device-scanner currently uses rpkg to build everything (including source).

This makes the spec heavier and slower than it needs to be in terms of building, as everything is pulled in and done inside a mock container.

Ideally, we should be able to use the make srpm option as that would allow us to rebuild entirely from the Copr interface. Unfortunately the chroot is pinned to a newer fedora and does not include a way to alter it's provided mock config, so it's not an option (track: https://pagure.io/copr/copr/issue/310).

That leaves us with the option to use module-tools to build locally and push changes via copr-cli, which is what this patch will implement.

Signed-off-by: Joe Grund <joe.grund@intel.com>