Remove condition on VDev using whole_disk
We currently predicate using `whole_disk` to include VDevs in the list
of matched devices.

However, when using /dev/mapper devices the whole_disk flag is not set
to true (see https://github.com/zfsonlinux/zfs/issues/2296).

Due to the unreliability of this method + how we are now accounting for
pool partitions, we should be able to remove this check.

Signed-off-by: Joe Grund <joe.grund@intel.com>