Remove vg_size check
We currently have a vg_size check as a udev
script. This appears to be causing a lockup
in my testing.

Remove the vg_size check from within udev and
zero out the result.

We don't display the size anywhere directly in IML
and once we model device types as a DU on the device-scanner
side we can get the size via an async call.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>