Normalize ZFS paths to devices
We are currently doing a naive match to resolve zpools
to backing devices on other nodes. This makes the assumption
that the chosen `Uevent.path` is equivalent to the `VDev.path`
entered by the user when creating the pool, and does not hold
when the user enters a different path.

Normalize any given `VDev.path` to a by-id/* path, which should
be consistent between nodes on a particular tick.

In addition clean up the munging / parsing code and remove a
bunch of mutable state.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>