Update device-scanner
device-scanner was using a mpsc for connection handling.

While this was a nice approach, it ended up blocking
all connections until the previous ones had finished writing.

Instead, use a custom Future that holds a socket and a mpsc
and hooks the mpsc up to the socket.

By doing this, writes will no longer block and things can move in
parallel.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>