Share Devices
Fixes #247

Devices such as mdraid, volume groups, and zpools may be shared between hosts.

If a device is shared it means:

All it's underlying devices are also shared (ex: VDev tree for a zpool).
There are no node-local devices built on the device.
In order to determine if a device is shared between hosts, we can look at the SCSI id + the type of the device.

We can then use that to build a cluster-wide DAG from hosts -> devices.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>