Use single-threaded runtime for zed-enhancer daemon
ZED events are emitted by a single thread, however the zed-enhancer
daemon is multithreaded. This means if there are multiple incoming ZED
events in a short time span, they may finish out of order.

Switch to using the single threaded runtime for zed-enhancer so events
are guaranteed to be processed serially.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>