Add shared VGs and LVs
When a VG is present in a device tree, we can tell if it can be used on
another host by inspecting the underlying PVs.

If all PVs are present on another host, then we can presume the VG can
be available on that host as well.

Add a fn that will look at the PVs and add VGs / LVs to other hosts when
they match.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>